// frontend/types.ts
export type ReserveDetail = {
  table_id: number;
  date: string;
  time_slot: string;
  guest_count: number;
  customer_name: string;
  contact: string;
};

export type BookingSummary = {
  table_id: number;
  date: string;
  time_slot: string;
  guest_count: number;
  customer_name: string;
  contact: string;
};